fix(deps): dev-preview.21

+++ src/register.ts
@@ -1,48 +1,58 @@
 import {
-  DeleteAction,
   DocumentActionComponent,
   DocumentActionDescription,
   DocumentActionProps,
-  DocumentBadgeComponent,
-  PublishAction
-} from 'sanity/desk'
+  DocumentBadgeComponent
+} from 'sanity'
 import {ScheduledBadge} from './badges/ScheduledBadge'
 import {createScheduleAction, createUnScheduleAction} from './actions/schedule'
 import {schedulingEnabled, useScheduleMetadata} from './scheduling'
 import {TypeConfig} from './types'
 
-export function CalendarDeleteAction(
-  params: DocumentActionProps
-): DocumentActionDescription | null {
-  const metadata = useScheduleMetadata(params.id)
+export const createCalendarDeleteAction = (actions: DocumentActionComponent[]) =>
+  function CalendarDeleteAction(params: DocumentActionProps): DocumentActionDescription | null {
+    const DefaultDelete = actions.find(a => a.action === 'delete')
+    if (!DefaultDelete) {
+      throw new Error('actions did not contain any actions with action === "delete"')
+    }
 
-  const onComplete = () => {
-    metadata.delete()
-    params.onComplete()
+    const metadata = useScheduleMetadata(params.id)
+
+    const onComplete = () => {
+      metadata.delete()
+      params.onComplete()
+    }
+
+    return DefaultDelete({
+      ...params,
+      onComplete
+    })
   }
 
-  return DeleteAction({
-    ...params,
-    onComplete
-  })
-}
+export const createCalendarPublishAction = (actions: DocumentActionComponent[]) =>
+  function CalendarPublishAction(params: DocumentActionProps): DocumentActionDescription | null {
+    const DefaultPublish = actions.find(a => a.action === 'publish')
+    if (!DefaultPublish) {
+      throw new Error('actions did not contain any actions with action === "publish"')
+    }
 
-export function CalendarPublishAction(
-  params: DocumentActionProps
-): DocumentActionDescription | null {
-  const metadata = useScheduleMetadata(params.id)
+    const metadata = useScheduleMetadata(params.id)
 
-  const result = PublishAction(params) as DocumentActionDescription
+    const result = DefaultPublish(params)
 
-  return {
-    ...result,
-    onHandle: () => {
-      // eslint-disable-next-line no-unused-expressions
-      result?.onHandle && result.onHandle()
-      metadata.delete()
+    if (!result) {
+      throw new Error('action with action === "publish" returned null')
+    }
+
+    return {
+      ...result,
+      onHandle: () => {
+        // eslint-disable-next-line no-unused-expressions
+        result?.onHandle && result.onHandle()
+        metadata.delete()
+      }
     }
   }
-}
 
 export function addActions(
   {type, types}: {type: string; types: TypeConfig[]},
@@ -52,12 +62,12 @@ export function addActions(
     const pluginActions: DocumentActionComponent[] = [
       createScheduleAction(types),
       createUnScheduleAction(types),
-      CalendarPublishAction,
-      CalendarDeleteAction
+      createCalendarPublishAction(actions),
+      createCalendarDeleteAction(actions)
     ]
 
     const defaultActions = actions.filter(action => {
-      return action !== DeleteAction && action !== PublishAction
+      return action.action !== 'delete' && action.action !== 'publish'
     })
 
     return [...pluginActions, ...defaultActions]
